+++ common/.zplugins/sudo.plugin.zsh
@@ -1,9 +1,8 @@
-# https://github.com/robbyrussell/oh-my-zsh/tree/master/plugins/sudo
 # ------------------------------------------------------------------------------
 # Description
 # -----------
 #
-# sudo will be inserted before the command
+# sudo or sudoedit will be inserted before the command
 #
 # ------------------------------------------------------------------------------
 # Authors
@@ -15,7 +14,17 @@
 
 sudo-command-line() {
     [[ -z $BUFFER ]] && zle up-history
-    [[ $BUFFER != sudo\ * ]] && LBUFFER="sudo $LBUFFER"
+    if [[ $BUFFER == sudo\ * ]]; then
+        LBUFFER="${LBUFFER#sudo }"
+    elif [[ $BUFFER == $EDITOR\ * ]]; then
+        LBUFFER="${LBUFFER#$EDITOR }"
+        LBUFFER="sudoedit $LBUFFER"
+    elif [[ $BUFFER == sudoedit\ * ]]; then
+        LBUFFER="${LBUFFER#sudoedit }"
+        LBUFFER="$EDITOR $LBUFFER"
+    else
+        LBUFFER="sudo $LBUFFER"
+    fi
 }
 zle -N sudo-command-line
 # Defined shortcut keys: [Esc] [Esc]