+++ common/.zplugins/zsh-autosuggestions/spec/strategies/match_prev_cmd_spec.rb
@@ -1,4 +1,6 @@
-describe 'the match_prev_cmd strategy' do
+require 'strategies/special_characters_helper'
+
+describe 'the `match_prev_cmd` strategy' do
   let(:options) { ['ZSH_AUTOSUGGEST_STRATEGY=match_prev_cmd'] }
 
   it 'suggests the last matching history entry after the previous command' do
@@ -14,4 +16,6 @@ describe 'the match_prev_cmd strategy' do
       wait_for { session.content }.to eq('ls bar')
     end
   end
+
+  include_examples 'special characters'
 end